requirements.txt
fastapi
uvicorn
pandas
joblib
python-multipart
google-generativeai
